Se añadió la interfaz de la pagina eliminarcuento.html y se hizo responsiva la pagina

+++ eliminarCuento.js
@@ -3,23 +3,22 @@ const db = firebase.firestore();
 const getCuentosFabula = () => db.collection('Fabulas').get();
 const getCuentosClasicos = () => db.collection('Clasicos').get();
 const getCuentosFantasia = () => db.collection('Fantasia').get();
-
+ 
+var titulospan = document.getElementById('text1'); 
 botonEliminar.onclick = async(e) =>{
       const coleccionFabulas = await getCuentosFabula();
       const coleccionFantasia = await getCuentosFantasia();
       const coleccionClasicos = await getCuentosClasicos();
       cuentoEliminado = document.getElementById('cuentobox').value;
-      var categoria = document.getElementById("categorias").value;
-       
-        botonEliminar = document.getElementById("botonEliminar");
-        var existe=false;
-        if(categoria=="Fantasia"){
-            
-        coleccionFantasia.forEach(doc =>{
+      var categoria = document.getElementById('categorias').value;
+        botonEliminar = document.getElementById('botonEliminar');
+        console.log(cuentoEliminado);
+        if(cuentoEliminado!= ""){
+           if(categoria=="Fantasia"){
+            coleccionFantasia.forEach(doc =>{
              if (doc.data().TituloCuento==cuentoEliminado){
                 doc.ref.delete();
-                alert("Cuento borrado de la Base de Datos");
-                existe=true;
+                alert("Cuento borrado de la categoría Fantasía");
              }
         });
         }else{
@@ -27,8 +26,7 @@ botonEliminar.onclick = async(e) =>{
             coleccionClasicos.forEach(doc =>{
                  if (doc.data().TituloCuento==cuentoEliminado){
                     doc.ref.delete();
-                alert("Cuento borrado de la Base de Datos");
-                existe=true;
+                    alert("Cuento borrado de la categoría Clásicos");
                  }
             });
             }else{
@@ -36,16 +34,15 @@ botonEliminar.onclick = async(e) =>{
                 coleccionFabulas.forEach(doc =>{
                      if (doc.data().TituloCuento==cuentoEliminado){
                         doc.ref.delete();
-                        alert("Cuento borrado de la Base de Datos");
-                        existe=true;
+                        alert("Cuento borrado de la categoría Fábulas");
                      }
                  })
               }
             }
          }
-         if(existe==false){
-            alert("El cuento no existe en la colección");
-         }
-         document.getElementById("form").reset();
-}
-    
+          
+                document.getElementById("form").reset();
+        }else{
+            titulospan.innerHTML="Ingrese titulo del cuento";
+        }
+    }
